Add showing of yardcam images
Yardcam images are taken with a webcam and stored externally. An image
corresponding to an observation can now be shown in conjunction.
An image can be shown in conjunction with many observations as images
are taken more seldom than observations are recorded. The base path
for images is configured in config.edn.